move panicwrap handler to logging package
